requirements.txt
appnope==0.1.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
bottle==0.12.16
bottle-websocket==0.2.9
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.3.2
defusedxml==0.5.0
Eel==0.10.4
entrypoints==0.3
future==0.17.1
gevent==1.4.0
gevent-websocket==0.10.1
greenlet==0.4.15
idna==2.8
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itchat==1.3.10
jedi==0.13.3
Jinja2==2.10
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.0.1
MarkupSafe==1.1.1
matplotlib==3.0.2
mistune==0.8.4
nbconvert==5.4.1
nbformat==4.4.0
NetEaseMusicApi==1.0.3
notebook==5.7.5
numpy==1.16.2
pandocfilters==1.4.2
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.7.0
psutil==5.5.1
ptyprocess==0.6.0
Pygments==2.3.1
PyMySQL==0.9.3
pyparsing==2.3.1
pypng==0.0.19
PyQRCode==1.2.1
pyrsistent==0.14.11
python-dateutil==2.8.0
python-decouple==3.1
pyzmq==18.0.1
qtconsole==4.4.3
requests==2.21.0
scikit-learn==0.20.2
scipy==1.2.1
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
SQLAlchemy==1.3.1
tensorboardX==1.6
terminado==0.8.1
testpath==0.4.2
torch==1.0.1.post2
torchvision==0.2.2.post3
tornado==6.0.1
tqdm==4.31.1
traitlets==4.3.2
urllib3==1.24.1
utils==0.9.0
Wand==0.5.4
wcwidth==0.1.7
webencodings==0.5.1
whichcraft==0.5.2
widgetsnbextension==3.4.2
